feat(safari-13)!: Renamed safari-13-0 emulator to safari-13
