+++ emulator-plugins/emulate-safari-13/CHANGELOG.md
@@ -5,4 +5,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # 1.0.0-alpha.0 (2020-06-27)
 
-**Note:** Version bump only for package @secret-agent/emulate-safari-13-0
+**Note:** Version bump only for package @secret-agent/emulate-safari-13
