+++ emulator-plugins/emulate-safari-13/navigator.json
