+++ emulators/index.ts
@@ -4,7 +4,7 @@ import Utils from './lib/Utils';
 import UserAgents from './lib/UserAgents';
 import Chrome79 from '@secret-agent/emulate-chrome-79';
 import Chrome80 from '@secret-agent/emulate-chrome-80';
-import Safari13_0 from '@secret-agent/emulate-safari-13-0';
+import Safari13 from '@secret-agent/emulate-safari-13';
 import Browsers from './data/browsers.json';
 import Log from '@secret-agent/commons/Logger';
 
@@ -54,4 +54,4 @@ export { EmulatorPlugin, EmulatorPluginStatics, Utils, UserAgents };
 
 Emulators.load(Chrome80);
 Emulators.load(Chrome79);
-Emulators.load(Safari13_0);
+Emulators.load(Safari13);
